LPS-115161 Refactor AddPanelContext
